3175: rename map to mapcomponent to not shadow the global "Map" property
